Split GraphicsImpl into its own cpp file (#612)
* Split GraphicsImpl into its own cpp file. Also created platform-specific headers to avoid ifdefs.

* Moved more platform-specific functionality to platform-specific headers.

* Removed encapsulating platform specific headers folder.

* Forgot to change CMake.
