+++ Core/Graphics/CMakeLists.txt
@@ -1,15 +1,18 @@
 set(SOURCES
     "Include/Babylon/Graphics.h"
+    "Source/${BABYLON_NATIVE_PLATFORM}/GraphicsPlatform.h"
     "Source/BgfxCallback.cpp"
     "Source/BgfxCallback.h"
     "Source/Graphics.cpp"
+    "Source/GraphicsImpl.cpp"
     "Source/GraphicsImpl.h")
 
 add_library(Graphics ${SOURCES})
 warnings_as_errors(Graphics)
 
 target_include_directories(Graphics
-    PUBLIC "Include")
+    PUBLIC "Include"
+    PRIVATE "Source/${BABYLON_NATIVE_PLATFORM}")
 
 target_link_to_dependencies(Graphics
     PUBLIC JsRuntime
@@ -25,7 +28,9 @@ set_property(TARGET Graphics PROPERTY FOLDER Core)
 source_group(TREE ${CMAKE_CURRENT_SOURCE_DIR} FILES ${SOURCES})
 
 add_library(GraphicsInternal INTERFACE)
-target_include_directories(GraphicsInternal INTERFACE "Source")
+target_include_directories(GraphicsInternal
+    INTERFACE "Source"
+    INTERFACE "Source/${BABYLON_NATIVE_PLATFORM}")
 target_link_to_dependencies(GraphicsInternal
     INTERFACE Graphics
     INTERFACE JsRuntimeInternal
