+++ Core/Graphics/Source/Apple/GraphicsPlatform.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <bgfx/bgfx.h>
+
+// On Apple devices we need to set BGFX_RESET_FLIP_AFTER_RENDER so that draw calls are immediately pushed
+// to the Metal command buffer to avoid a one frame delay.
+#define BGFX_RESET_FLAGS (BGFX_RESET_VSYNC | BGFX_RESET_MSAA_X4 | BGFX_RESET_MAXANISOTROPY | BGFX_RESET_FLIP_AFTER_RENDER)
+
+namespace Babylon
+{
+    constexpr auto inline BgfxDefaultRendererType{bgfx::RendererType::Metal};
+}
