+++ Core/Graphics/Source/UWP/GraphicsPlatform.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <bgfx/bgfx.h>
+
+#define BGFX_RESET_FLAGS (BGFX_RESET_VSYNC | BGFX_RESET_MSAA_X4 | BGFX_RESET_MAXANISOTROPY)
+
+namespace Babylon
+{
+    constexpr auto inline BgfxDefaultRendererType{bgfx::RendererType::Direct3D11};
+}
